+++ backend/src/main/java/com/example/backend/Ports/In/VentaProductoIn.java
@@ -7,8 +7,8 @@ import com.example.backend.Models.VentaProducto;
 
 public interface VentaProductoIn {
   VentaProducto crearVentaProducto(VentaProducto ventaProducto);
-  Optional<VentaProducto> obtenerVentaProductoPorId(Integer id);
+  Optional<VentaProducto> obtenerVentaProductoPorId(Long id);
   List<VentaProducto> listarVentaProductos();
-  VentaProducto actualizarVentaProducto(Integer id, VentaProducto ventaProducto);
-  void eliminarVentaProducto(Integer id);
+  VentaProducto actualizarVentaProducto(Long id, VentaProducto ventaProducto);
+  void eliminarVentaProducto(Long id);
 }
